requirements.txt
aioredis==1.2.0
asgiref==2.3.2
asn1crypto==0.24.0
async-timeout==3.0.1
attrs==19.1.0
autobahn==19.3.3
Automat==0.7.0
certifi==2019.3.9
cffi==1.12.2
channels==2.1.7
channels-redis==2.3.3
chardet==3.0.4
constantly==15.1.0
cryptography==2.6.1
daphne==2.2.5
dj-database-url==0.5.0
Django==2.2.13
django-configurations==2.1
django-cors-headers==2.5.2
django-froala-editor==3.0.1
django-haystack==2.8.1
django-nine==0.1.13
django-rest-framework==0.1.0
django-storages==1.7.1
django-taggit==1.1.0
django-taggit-serializer==0.1.7
django-tagulous==0.14.0
djangorestframework==3.9.2
drf-extensions==0.4.0
drf-haystack==1.8.4
elasticsearch==2.0.0
hiredis==1.0.0
hyperlink==18.0.0
idna==2.8
incremental==17.5.0
ipaddress==1.0.22
isort==4.3.17
msgpack==0.6.1
psycopg2-binary==2.7.7
pycparser==2.19
PyHamcrest==1.9.0
PyJWT==1.7.1
python-dateutil==2.8.0
pytz==2018.9
requests==2.21.0
six==1.12.0
Twisted==18.9.0
txaio==18.8.1
Unipath==1.1
urllib3==1.24.1
zope.interface==4.6.0
boto3==1.9.153
Pillow==6.0.0
gunicorn==19.9.0
python-memcached==1.59
nose==1.3.7
django-nose
coreapi
python-slugify
